Move session related code to its own controller
